function-direct-upload/requirements.txt
functions-framework
google-cloud-storage
werkzeug